re-organize repo structure (#21)
* mv defaults/schemas to src/compass dir

+ move defaults and schemas folders to src/compass dir

+ rename the runconfig yaml files:
   - cslc_s1.yaml --> s1_cslc_radar.yaml
   - geo_cslc_s1.yaml --> s1_cslc_geo.yaml
   - update their usage in other scripts

+ update MANIFEST.in for the new yaml file locations

+ utils.helpers: use os.path.dirname(compass.__file__) to locate data files to replace the previous relative location using __file__ of the python script itself, to avoid troubles during profiling via scalene.

* rename geo/radar_grid.py

+ reference_radar_grid --> radar_grid.py

+ geogrid.py --> geo_grid.py

* rename s1_resample_burst to s1_resample

+ rename s1_resample_burst.py to s1_resample.py

+ del utils/unwrap_namespace.py, as it's duplicated to the wrap_namespace.unwrap_to_dict() and not used.

* rename geo_cslc to s1_geocode_slc

+ rename geo_cslc.py to s1_geocode_slc.py and update its usage

+ rename geo_cslc_metadata to geo_metadata

+ fix typos in journal.info() calls by adding the missing "s1_" prefix

* consistent parser/runconfig/log/date_str names

+ use consistent short parser var name: parser

+ use consistent runconfig var name: cfg

+ consistent journal output msg: same as script name

+ print/log time info without microseconds

+ s1_cslc_radar workflow: use YYYYMMDD as date_str w/o hyphen, to be shorter and consistent with geocode_slc workflow

* s1_cslc: add main() for entry_points

+ s1_geocode_slc: move the save metadata part from __main__ into run().

+ s1_cslc: add a main() to:
   - read the yaml file content
   - identify the geo / radar workflow based on processing steps
   - then call the corresponding workflow

+ setup.cfg: add entry_points for s1_cslc.py

+ README: update command line usage

* add helpers.get_module_name

* s1_geocode_slc: del for loop as it's only one burst

* set load_from_yaml(workflow_name='s1_cslc_geo/radar')

* refactored yaml_argparse to accommodate new arguments
refactored s1_cslc to allow for grid type

* s1_cslc: add run() to facilitate ext function call
