+++ MANIFEST.in
@@ -1,1 +1,1 @@
-recursive-include src/compass/utils/* *.yaml
+recursive-include src/compass/* *.yaml
