+++ setup.cfg
@@ -24,3 +24,8 @@ install_requires =
 
 [options.packages.find]
 where = src
+
+[options.entry_points]
+console_scripts =
+    s1_cslc.py = compass.s1_cslc:main
+
